3-implement-trivia-routes (#31)
* Creating MD file

Created MD file for initial JSON design choice. This commit is also an initial test to check branch

* updated JSON structure

* Successfully returned JSON

We tested a JSON response and received what we expected. We also wrote a database query statement and created a response dictionary that we will uncomment once database has data

* Trivia specific number route returns expected response

We added the dictionary that will be jsonified and made the route return the jsonified item.
We also added the error response when there are no facts for the specific number

* No changes made

* Addressed when error message should appear

We altered the condition for the if statement on error messages so that when result is an empty list, an error message would appear as well

* Began creating test data

We created test data, but currently have issue with the Blueprint and Blueprint not having app.Working through to get initial tests running

* Refactored seed file to python3

previous method was deprecated and throwing errors

* Finished functionality for both trivia/number and trivia/random routes

* Added tests for trivia/number and trivia/rnadom routes

* Refactored per PR requested changes

We changed the variable names. Edited any indentation issues. Added to docstrings.

* Deleted this file

* Made format changes

Co-authored-by: Ezra Chung <ezra.chung97@gmail.com>