7 implement date routes (#33)
* Added validation

* Created Tests

* Created test

Created test for new class method that will take the day of year and convert to tuple of (month,day)

* Added new class method

Purpose is for the JSON that will be returned with the month and day. We choose to do this on the models instead of in the route to keep the level of truthfulness. -wip

* Date Mode: Finished date_from_day_of_year class method

Finished logic to convert day of year to month and day

* Date: models

Refactored names for easier readability

* Date routes: Implemented date routes (both specific and random

All tests pass

* Date Model: Moved doctests into unittests

Doctests were giving metadata issues (need to investigate further)

* Date routes: refactored naming

month_fact variable name was slightly confusing about what it contained so changed it to month_num
same with day_fact

* Date test_models: Moved doctests from class model into unittest

Doctests were giving metadata issues (need to investigate further)

* Dates models: Added data validation

Added type and range of data validation

* Dates routes: Added error handling

Added error handling in routes because we added data validation to Date class methods

* Dates model tests: Added tests for error handling

Checked to ensure correct error type and error messages are being thrown

* Dates routes: added testing for data validation

Added testing to check we received appropriate error responses when month and day were integers but not in range of normal month or day numbers

* made small tweaks in prep for merge to main

Co-authored-by: Ezra Chung <ezra.chung97@gmail.com>
Co-authored-by: Brittiany Juravic <brittiany1212@gmail.com>